requirements.txt
googlemaps==4.10.0
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
tldextract==5.1.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
geopy==2.4.1
beautifulsoup4==4.12.3
numpy==1.26.4
geographiclib==2.0 